requirements.txt
django
psycopg2-binary
pillow
django-widget-tweaks
djangorestframework